refactor: create list modal
