+++ apps/renderer/src/modules/settings/tabs/lists/modals.tsx
@@ -24,8 +24,7 @@ import {
   FormMessage,
 } from "~/components/ui/form"
 import { Input } from "~/components/ui/input"
-import { LoadingCircle } from "~/components/ui/loading"
-import { useModalStack } from "~/components/ui/modal"
+import { useCurrentModal } from "~/components/ui/modal"
 import { ScrollArea } from "~/components/ui/scroll-area"
 import {
   Table,
@@ -35,15 +34,12 @@ import {
   TableHeader,
   TableRow,
 } from "~/components/ui/table"
-import { Tooltip, TooltipContent, TooltipPortal, TooltipTrigger } from "~/components/ui/tooltip"
 import { views } from "~/constants"
-import { useAuthQuery, useI18n } from "~/hooks/common"
 import { apiClient } from "~/lib/api-fetch"
 import { createErrorToaster } from "~/lib/error-parser"
-import { cn, isBizId } from "~/lib/utils"
+import { isBizId } from "~/lib/utils"
 import type { FeedModel } from "~/models"
 import { ViewSelectorRadioGroup } from "~/modules/shared/ViewSelectorRadioGroup"
-import { Balance } from "~/modules/wallet/balance"
 import { Queries } from "~/queries"
 import {
   getFeedById,
@@ -54,143 +50,6 @@ import {
 import { useListById } from "~/store/list"
 import { subscriptionActions, useSubscriptionStore } from "~/store/subscription"
 
-export const SettingLists = () => {
-  const t = useI18n()
-  const listList = useAuthQuery(Queries.lists.list())
-
-  const { present } = useModalStack()
-
-  return (
-    <section className="mt-4">
-      <div className="mb-4 space-y-2 text-sm">
-        <p>{t.settings("lists.info")}</p>
-      </div>
-      <Button
-        onClick={() => {
-          present({
-            title: t.settings("lists.create"),
-            content: ({ dismiss }) => <ListCreationModalContent dismiss={dismiss} />,
-          })
-        }}
-      >
-        <i className="i-mgc-add-cute-re mr-1 text-base" />
-        {t.settings("lists.create")}
-      </Button>
-      <Divider className="mb-6 mt-8" />
-      <div className="flex flex-1 flex-col">
-        <ScrollArea.ScrollArea viewportClassName="max-h-[380px]">
-          {listList.data?.length ? (
-            <Table className="mt-4">
-              <TableHeader className="border-b">
-                <TableRow className="[&_*]:!font-semibold">
-                  <TableHead size="sm">{t.settings("lists.title")}</TableHead>
-                  <TableHead size="sm">{t.settings("lists.view")}</TableHead>
-                  <TableHead size="sm">{t.settings("lists.fee.label")}</TableHead>
-                  <TableHead size="sm">{t.settings("lists.subscriptions")}</TableHead>
-                  <TableHead size="sm">{t.settings("lists.earnings")}</TableHead>
-                  <TableHead size="sm" className="center">
-                    {t.common("words.actions")}
-                  </TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody className="border-t-[12px] border-transparent [&_td]:!px-3">
-                {listList.data?.map((row) => (
-                  <TableRow key={row.title} className="h-8">
-                    <TableCell size="sm">
-                      <a
-                        target="_blank"
-                        href={`${WEB_URL}/list/${row.id}`}
-                        className="inline-flex items-center gap-2 font-semibold"
-                      >
-                        {row.image && (
-                          <Avatar className="size-6">
-                            <AvatarImage src={row.image} />
-                          </Avatar>
-                        )}
-                        <span className="inline-block max-w-[200px] truncate">{row.title}</span>
-                      </a>
-                    </TableCell>
-                    <TableCell size="sm">
-                      <Tooltip>
-                        <TooltipTrigger asChild>
-                          <span
-                            className={cn("inline-flex items-center", views[row.view].className)}
-                          >
-                            {views[row.view].icon}
-                          </span>
-                        </TooltipTrigger>
-                        <TooltipPortal>
-                          <TooltipContent>{t(views[row.view].name)}</TooltipContent>
-                        </TooltipPortal>
-                      </Tooltip>
-                    </TableCell>
-                    <TableCell size="sm">
-                      <div className="flex items-center gap-1">
-                        {row.fee}
-                        <i className="i-mgc-power shrink-0 text-lg text-accent" />
-                      </div>
-                    </TableCell>
-                    <TableCell size="sm">{row.subscriptionCount}</TableCell>
-                    <TableCell size="sm">
-                      <Balance>{BigInt(row.purchaseAmount || 0n)}</Balance>
-                    </TableCell>
-                    <TableCell size="sm" className="center">
-                      <Tooltip>
-                        <TooltipTrigger asChild>
-                          <Button
-                            variant="ghost"
-                            onClick={() => {
-                              present({
-                                title: t.settings("lists.feeds.manage"),
-                                content: () => <ListFeedsModalContent id={row.id} />,
-                              })
-                            }}
-                          >
-                            <i className="i-mgc-inbox-cute-re" />
-                          </Button>
-                        </TooltipTrigger>
-                        <TooltipPortal>
-                          <TooltipContent>{t.common("words.manage")}</TooltipContent>
-                        </TooltipPortal>
-                      </Tooltip>
-                      <Tooltip>
-                        <TooltipTrigger asChild>
-                          <Button
-                            variant="ghost"
-                            onClick={() => {
-                              present({
-                                title: t.settings("lists.edit.label"),
-                                content: ({ dismiss }) => (
-                                  <ListCreationModalContent dismiss={dismiss} id={row.id} />
-                                ),
-                              })
-                            }}
-                          >
-                            <i className="i-mgc-edit-cute-re" />
-                          </Button>
-                        </TooltipTrigger>
-                        <TooltipPortal>
-                          <TooltipContent>{t.common("words.edit")}</TooltipContent>
-                        </TooltipPortal>
-                      </Tooltip>
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-          ) : listList.isLoading ? (
-            <LoadingCircle size="large" className="center absolute inset-0" />
-          ) : (
-            <div className="mt-36 w-full text-center text-sm text-zinc-400">
-              <p>{t.settings("lists.noLists")}</p>
-            </div>
-          )}
-        </ScrollArea.ScrollArea>
-      </div>
-    </section>
-  )
-}
-
 const formSchema = z.object({
   view: z.string(),
   title: z.string().min(1),
@@ -199,7 +58,8 @@ const formSchema = z.object({
   fee: z.number().min(0),
 })
 
-const ListCreationModalContent = ({ dismiss, id }: { dismiss: () => void; id?: string }) => {
+export const ListCreationModalContent = ({ id }: { id?: string }) => {
+  const { dismiss } = useCurrentModal()
   const { t } = useTranslation(["settings", "common"])
 
   const list = useListById(id)
